Bugfix: asset
Fixed exception that would be thrown when adding asset to city.assets
within a for-each loop.  New assets are now added to city.newAssets, and
the entire list added to city.assets after the for-each loop.

+++ ASkyFullOfStars/Globals/valve.vb
@@ -21,7 +21,7 @@ Module valve
         Return total
     End Function
 
-    Public Function sign(value As Integer) As String
+    Public Function sign(value As Decimal) As String
         If value < 0 Then Return "" Else Return "+"
     End Function
